Move source types into source module
